move ticket form layout into tests directory
